+++ src/pages/contact-us.js
@@ -1,0 +1,11 @@
+import React from "react";
+const ContactUsPage = () => (
+	<div>
+		<h1> Contact Us</h1>
+		<h2>Content Change</h2>
+		For any concerns regarding content please contact us at
+		<a href="mailto:ausgov@digital.gov.au"> ausgov@digital.gov.au</a>
+		<br />
+	</div>
+);
+export default ContactUsPage;